backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.115.12",
    "google>=3.0.0",
    "google-genai>=1.19.0",
    "ipykernel>=6.29.5",
    "kernel>=0.5.0",
    "ollama>=0.5.1",
    "openai>=1.84.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.9.0",
]
